0.2 Beta
Major revamp, now with log panel.

+++ src/org/orecraft/budjet/PropertyLoader.java
@@ -5,14 +5,18 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.Properties;
 
+import javax.swing.filechooser.FileSystemView;
+
 public class PropertyLoader {
 	
 	public void refreshProperties(String newFirst, String newLast, double newBalance, int newDepositAmount, int newWithdrawAmount)
 	{
 		Properties properties = new Properties();
 		try {
-			File file = new File("BudJet.txt");
-			if(file.exists())
+			FileSystemView fw = FileSystemView.getFileSystemView();
+			File folder = new File(fw.getHomeDirectory(), "/Library/Application Support/BudJet");
+			File file = new File(fw.getHomeDirectory(), "/Library/Application Support/BudJet/BudJetData.txt");
+			if(folder.exists() && file.exists())
 			{
 				FileOutputStream fileoutputstream = new FileOutputStream(file);
 				properties.setProperty("firstName", newFirst);
@@ -24,9 +28,10 @@ public class PropertyLoader {
 				fileoutputstream.close();
 				System.out.println("Data has been updated.");
 			}
-			else {
-				file.delete();
+			else if(folder.exists() && !file.exists())
+			{
 				file.createNewFile();
+				System.out.println("Created database.");
 				FileOutputStream fileoutputstream = new FileOutputStream(file);
 				properties.setProperty("firstName", "User");
 				properties.setProperty("lastName", "");
@@ -35,12 +40,29 @@ public class PropertyLoader {
 				properties.setProperty("withdrawAmount", Integer.toString(0));
 				properties.store(fileoutputstream, "Main BudJet data.");
 				fileoutputstream.close();
-				System.out.println("BudJet.txt file has been created.");
+			}
+			else if(!folder.exists() && !file.exists())
+			{
+				folder.mkdir();
+				System.out.println("Created directory.");
+				file.createNewFile();
+				System.out.println("Created database.");
+				FileOutputStream fileoutputstream = new FileOutputStream(file);
+				properties.setProperty("firstName", "User");
+				properties.setProperty("lastName", "");
+				properties.setProperty("balance", Double.toString(0.00));
+				properties.setProperty("depositAmount", Integer.toString(0));
+				properties.setProperty("withdrawAmount", Integer.toString(0));
+				properties.store(fileoutputstream, "Main BudJet data.");
+				fileoutputstream.close();
+			}
+			else {
+				System.out.println("An error occured.");
 			}
 		}
 		catch (IOException e)
 		{
-			System.out.println("An error occured while updating database.");
+			System.err.println("An error occured while updating database.");
 		}
 	}
 }
